Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -6,9 +6,9 @@ char* welcome() {
 
 int main() {
     int a,b,c;
-    double avg;
+    float avg;
     scanf("%d %d %d",&a,&b,&c);
     avg=(a+b+c)/3;
-    printf("Average: %d\n",avg);
+    printf("Average: %.2f\n",avg);
 
 }
